#poc_v3: refactor fixtures
