+++ contracts_arbitrum/contracts/v3/Controller.sol
@@ -1,20 +1,44 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts/token/ERC1155/ERC1155.sol";
 import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
 import "hardhat/console.sol";
 import "./SeriesMaster.sol";
-import "./AssetsToken.sol";
+import "./AssetsManager.sol";
 
 contract Controller is Initializable, OwnableUpgradeable {
-  AssetsToken _assetsManager;
+  AssetsManager _assetsManager;
   SeriesMaster _seriesMaster;
 
-  function initialize(address assetsManager, address seriesMaster) initializer public {
-    _assetsManager = ERC1155(assetsManager);
-    _seriesMaster = ERC1155(seriesMaster);
+  function initialize(address assetsManager, address payable seriesMaster) initializer public {
+    _assetsManager = AssetsManager(assetsManager);
+    _seriesMaster = SeriesMaster(seriesMaster);
     __Ownable_init();
+  }
+
+  function listAsset(
+    string memory name,
+    string memory symbol,
+    string memory title,
+    string memory description,
+    uint8 status,
+    uint256 tokenInfo_totalSupply,
+    uint256 tokenInfo_apr,
+    uint256 tokenInfo_tokenPriceDe6,
+    string memory propertyInfo_images
+  ) public {
+    _assetsManager.createAsset(
+      name,
+      symbol,
+      title,
+      description,
+      status,
+      tokenInfo_totalSupply,
+      tokenInfo_apr,
+      tokenInfo_tokenPriceDe6,
+      propertyInfo_images
+    );
+    _seriesMaster.createSeries(address(this), name);
   }
 }
