+++ contracts_arbitrum/testUtils/types.ts
@@ -1,0 +1,18 @@
+export enum AssetStatuses {
+  'upcoming' = 1,
+  'active'= 2,
+  'sold_out'= 3,
+  'disabled'= 4,
+}
+
+export type AssetInput = {
+  name: string,
+  symbol: string,
+  title: string,
+  description: string,
+  status: number,
+  tokenInfo_totalSupply: number,
+  tokenInfo_apr: number,
+  tokenInfo_tokenPriceDe6: number,
+  propertyInfo_images: string,
+}
